+++ src/components/MainSection/WeaponViewer/WeaponPanel/WeaponPanel.vue
@@ -125,6 +125,9 @@ function onPerkClicked(column: PerkColumnNumber) {
 .description {
     display: flex;
     flex-direction: column;
+    h1, h3 {
+        margin: 0;
+    }
 }
 .description h1 {
     font-size: 32px;