+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Description TBD
 * Added filters for seasonal weapons and various other collections (seasons prior to Season of Worthy not yet implemented).
 * Insights/community info on perks added to tooltips for perks and mods.
 * Glossary implemented to display list of all insights/community info.
+* Fixed background image placement.
+* Removed glossary terms for Rampage Spec, Dragonfly Spec, and Surrounded Spec.
 
 ## **Improve Page Load Times + Quality of Life** [ *1.1.0* ] 2023-02-13
 This change brings improvements to page load times as well as some quality of life changes such as